complete todo with crud

+++ src/components/AddTodo.js
@@ -1,16 +1,18 @@
 import React, { useEffect, useState } from 'react'
 import { Button, Table } from 'react-bootstrap'
-import { DeleteTodo, addTodoList, getTodoList } from '../action/Action'
+import { DeleteTodo, addTodoList, getTodoList, updateTodo } from '../action/Action'
 import { useDispatch, useSelector } from 'react-redux'
-export default function AddTodo() {
+export default function AddTodo({ handleEdit, formVisibility, editTodo }) {
 
     const dispatch = useDispatch();
     const todoList = useSelector(state => state.todo.todos)
-    // console.log(todoList, "uhuhyuhuhuhuhuhuhuhuh")
     const [formData, setFormData] = useState({
+        todoId: '',
         title: '',
         description: ''
     })
+
+    //get data
     const handleChange = (e) => {
         e.preventDefault()
         const { name, value } = e.target;
@@ -20,6 +22,7 @@ export default function AddTodo() {
         }))
     }
 
+    //add data
     const handleSubmit = async event => {
         event.preventDefault()
         const { title, description } = formData;
@@ -29,6 +32,7 @@ export default function AddTodo() {
             await dispatch(addTodoList(todoData));
             await dispatch(getTodoList());
             setFormData({
+                id: '',
                 title: '',
                 description: ''
             })
@@ -38,11 +42,35 @@ export default function AddTodo() {
         }
     }
 
-    const handleDelete = async (id) => {
+    //delete data
+    const handleDelete = (id) => {
+        if (window.confirm('Are you sure you want to delete this todo')) {
+            dispatch(DeleteTodo(id));
+            dispatch(getTodoList())
+        }
+    }
 
-        await dispatch(DeleteTodo(id));
-        await dispatch(getTodoList())
+    //update data
+    const handleSelectTodo = (todo) => {
+        setFormData({
+            todoId: todo._id,
+            title: todo.title,
+            description: todo.description
+        });
+        handleEdit(true);
+    }
+    const handleUpdate = (e) => {
+        e.preventDefault();
 
+        let editedObj = {
+            todoId: formData.todoId,
+            title: formData.title,
+            description: formData.description
+        }
+
+        dispatch(updateTodo(editedObj))
+        dispatch(getTodoList());
+        window.location.reload()
     }
     useEffect(() => {
         dispatch(getTodoList())
@@ -51,32 +79,60 @@ export default function AddTodo() {
     return (
         <>
             <center>
-                <form onSubmit={handleSubmit}>
-                    <input
-                        className='mb-3'
-                        type='text'
-                        name='title'
-                        placeholder='Enter Title'
-                        value={formData.title}
-                        onChange={handleChange}
-                        style={{ borderRadius: '5px' }}
-                    /> {' '}
-                    <input
-                        className='mb-3'
-                        type='text'
-                        name='description'
-                        placeholder='Enter Description'
-                        value={formData.description}
-                        onChange={handleChange}
-                        style={{ borderRadius: '5px' }}
-                    /> {' '}
-                    <Button
-                        type='submit'
-                        className='mb-1'
-                        size='sm'
-                    >Add Todo</Button>
-                </form>
-                <Table bordered striped hover size='sm' variant='dark' style={{ width: "600px" }}>
+                {formVisibility === false ?
+                    <form onSubmit={handleSubmit}>
+                        <input
+                            className='mb-3'
+                            type='text'
+                            name='title'
+                            placeholder='Enter Title'
+                            value={formData.title}
+                            onChange={handleChange}
+                            style={{ borderRadius: '5px' }}
+                        /> {' '}
+                        <input
+                            className='mb-3'
+                            type='text'
+                            name='description'
+                            placeholder='Enter Description'
+                            value={formData.description}
+                            onChange={handleChange}
+                            style={{ borderRadius: '5px' }}
+                        /> {' '}
+                        <Button
+                            type='submit'
+                            className='mb-1'
+                            size='sm'
+                        >Add Todo</Button>
+                    </form> : <form onSubmit={handleUpdate}>
+                        <input
+                            className='mb-3'
+                            type='text'
+                            name='title'
+                            value={formData.title}
+                            onChange={handleChange}
+                            placeholder='Update Title'
+                            style={{ borderRadius: '5px' }}
+                        /> {' '}
+                        <input
+                            className='mb-3'
+                            type='text'
+                            name='description'
+                            value={formData.description}
+                            onChange={handleChange}
+                            placeholder='Update Description'
+                            style={{ borderRadius: '5px' }}
+                        /> {' '}
+                        <Button
+                            type='submit'
+                            className='mb-1'
+                            size='sm'
+
+                        >Update Todo</Button>
+                    </form>
+                }
+
+                <Table bordered striped hover size='sm' variant='dark' style={{ width: "500px" }}>
                     <thead>
                         <tr>
                             <th>Title</th>
@@ -86,15 +142,20 @@ export default function AddTodo() {
                     </thead>
                     <tbody>
                         {todoList && todoList.length !== 0 ? (
-                            todoList?.data?.map(todo => (
+                            todoList.data?.map(todo => (
                                 <tr key={todo._id}>
                                     <td>{todo.title}</td>
                                     <td>{todo.description}</td>
                                     <td>
-                                        <tr className='d-flex justify-content-evenly'>
-                                            <td><Button type='submit'>Edit</Button></td>
-                                            <td><Button type='submit' variant='danger' onClick={() => { handleDelete(todo._id) }}>Delete</Button></td>
-                                        </tr>
+
+                                        {formVisibility === false && (
+                                            <>
+                                                <tr className='d-flex justify-content-evenly'>
+                                                    <td><Button type='submit' onClick={() => handleSelectTodo(todo, todo._id)}>Edit</Button></td>
+                                                    <td><Button type='submit' variant='danger' onClick={() => { handleDelete(todo._id) }}>Delete</Button></td>
+                                                </tr>
+                                            </>
+                                        )}
                                     </td>
                                 </tr>
                             ))
@@ -106,7 +167,7 @@ export default function AddTodo() {
 
                     </tbody>
                 </Table>
-            </center>
+            </center >
         </>
     )
 }
